Presentable Attachments
Added support for a simple presenter mode using a boolean tracker, 'Attachment::isPresentable($flag)'. This allows a developer to set content blocks to behave either for editing or presentation.

Whenever 'AttachmentAwareTrait::attachments()' is used, 'Attachment::isPresentable()' will be TRUE; the trait is usually employed when previewing content in the Admin (from the widget) or displaying content on the front-end, and never for editing.

Changes PHP:
- AttachmentWidget: Removed 'withBaseUrl()' and 'pathWithBaseUrl()';
- Attachment: Added 'base_url' setter/getter;
- Attachment: Added 'createAbsoluteUrl()' and 'isRelativeUri()';
- Attachment: Added 'resolveUrls()';
- Attachment: Added 'isPresentable()';
- File, Image: Added 'src()' aliases to return absolute URI;
- Added missing setter/getter for “thumbnail” property on Attachment;

Changes Mustache:
- Image: Replaced '{{thumbnail}}'/'{{file}}' with '{{src}}';
- File: Replaced '{{baseUrl}}{{file}}' with '{{src}}';
- Text: Replaced {{withBaseUrl}}' with 'Attachment::isPresentable()' and 'Attachment::resolveUrls()';