Fix storing and parsing of l10n/allowNull values
Fixes a conflict on properties that are multilingual (`l10n`) and nullable (`allowNull`) where a localized field (for example `name_fr`) would convert an empty string to `null`.

Upon loading that data from source, the property would ignore the localized `null` field and return only the non-null field. A translatable property should be returned as a whole.

Changed:
- Method `StorablePropertyTrait::storageVal()` to stringify `Translation` objects earlier.
- Method `StorablePropertyTrait::storageVal()` to return empty string instead of `null` for multilingual/nullable property value (specifically, it's language field).
- Method `StorablePropertyTrait::parseFromFlatData()` to assign an empty string for null/missing localizations.