feat(admin): Add validation message to TinyMCE on save
Add support for TinyMCE's Notifications Manager to provide validity feedback: we add an error notification when the form dispatches an `invalid` event.

Since TinyMCE does not properly support editors inside dialogs, notifications are positionned absolutely, outside of the modal, from the bottom of the DOM. This risks positioning them outside of the viewport. To work around this issue, a debounced `scroll` event is used to reposition any notifications.